+++ web-dev-challenge/src/App/App.css
@@ -43,4 +43,5 @@ input {
   top: 296px;
   background: #25273C;
   border-radius: 7px;
+  text-align: center;
 }
