+++ web-dev-challenge/src/App/App.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import './App.css';
-import ToDoList from './ToDoList';
+import ToDoList from '../ToDoList/ToDoList';
 
 class App extends Component {
   constructor() {
